# openapi/v1/authentication.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Acervo Authentication API",
    "version": "1.0.0",
    "description": "Canonical Bearer-token authentication contract for Acervo. Runtime API addresses are supplied by each deployment and are not part of this contract."
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Authenticate a user and issue a Sanctum Bearer token",
        "description": "Returns a Sanctum Personal Access Token in data.token. This contract does not define or issue a refresh token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or unverified account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidCredentialsError"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity has no valid institution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionRequiredError"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/profile": {
      "get": {
        "operationId": "profile",
        "summary": "Get the authenticated user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Revoke the current Sanctum access token",
        "description": "Revokes only the Bearer token used for this request. Other access tokens issued to the same user remain valid.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Institution": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "email",
          "institution",
          "roles"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "institution": {
            "$ref": "#/components/schemas/Institution"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "editor",
                "reader"
              ]
            }
          }
        }
      },
      "LoginData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "user",
          "token"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "minLength": 1,
            "description": "Sanctum Personal Access Token to send using the Bearer authentication scheme."
          }
        }
      },
      "ProfileData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "user"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "LoginSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "data",
          "message"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/LoginData"
          },
          "message": {
            "const": "Login successful."
          }
        }
      },
      "ProfileSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "data",
          "message"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/ProfileData"
          },
          "message": {
            "const": "Profile retrieved successfully."
          }
        }
      },
      "LogoutSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "data",
          "message"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {
            "type": "null"
          },
          "message": {
            "const": "Logout successful."
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "InvalidCredentialsError": {
        "$ref": "#/components/schemas/ErrorResponse",
        "examples": [
          {
            "success": false,
            "error": {
              "code": "AUTH_INVALID_CREDENTIALS",
              "message": "The provided credentials are invalid."
            }
          }
        ]
      },
      "InstitutionRequiredError": {
        "$ref": "#/components/schemas/ErrorResponse",
        "examples": [
          {
            "success": false,
            "error": {
              "code": "AUTH_INSTITUTION_REQUIRED",
              "message": "A valid institution is required to access the application."
            }
          }
        ]
      },
      "UnauthenticatedError": {
        "$ref": "#/components/schemas/ErrorResponse",
        "examples": [
          {
            "success": false,
            "error": {
              "code": "AUTH_UNAUTHENTICATED",
              "message": "Authentication is required."
            }
          }
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "fields"
            ],
            "properties": {
              "code": {
                "const": "VALIDATION_ERROR"
              },
              "message": {
                "const": "The given data was invalid."
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}